+++ code/patchwork/patchwork-core/src/traits/tree_policy_node.rs
@@ -7,7 +7,7 @@ pub trait TreePolicyNode {
     /// # Returns
     ///
     /// The number of times this node has been visited.
-    fn visit_count(&self) -> i32;
+    fn visit_count(&self) -> usize;
     /// The player whose turn it is at the current node.
     ///
     /// # Returns
@@ -34,7 +34,7 @@ pub trait TreePolicyNode {
     /// # Returns
     ///
     /// The maximum score of all games played from this node from the perspective of the given
-    fn maximum_score_for(&self, player: Self::Player) -> i32;
+    fn maximum_score_for(&self, player: Self::Player) -> f64;
     /// Gets the minimum score of all games played from this node from the perspective of the given
     /// player.
     ///
@@ -45,7 +45,7 @@ pub trait TreePolicyNode {
     /// # Returns
     ///
     /// The minimum score of all games played from this node from the perspective of the given
-    fn minimum_score_for(&self, player: Self::Player) -> i32;
+    fn minimum_score_for(&self, player: Self::Player) -> f64;
     /// Gets the difference between the maximum and minimum scores of the node.
     /// This is the same as the difference between the biggest win and biggest
     /// loss the node has had from the perspective of either player.
@@ -53,7 +53,7 @@ pub trait TreePolicyNode {
     /// # Returns
     ///
     /// The difference between the maximum and minimum scores of the node.
-    fn score_range(&self) -> i32 {
+    fn score_range(&self) -> f64 {
         let some_player = self.current_player();
         (self.maximum_score_for(some_player) - self.minimum_score_for(some_player)).abs()
     }
@@ -68,7 +68,7 @@ pub trait TreePolicyNode {
     ///
     /// The sum of the scores of all games played from this node from the perspective of the given
     /// player.
-    fn score_sum_for(&self, player: Self::Player) -> i64;
+    fn score_sum_for(&self, player: Self::Player) -> f64;
     /// Gets the average score of all games played from this node from the perspective of the given
     /// player.
     ///
@@ -84,7 +84,7 @@ pub trait TreePolicyNode {
         if self.visit_count() == 0 {
             return 0.0;
         }
-        self.score_sum_for(player) as f64 / self.visit_count() as f64
+        self.score_sum_for(player) / self.visit_count() as f64
     }
     /// Returns a prior believe of the value of the node if available.
     /// Otherwise returns 0.
